requirements.txt
affine==2.3.1
appnope==0.1.3
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
Cartopy==0.21.0
certifi==2022.9.24
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.0.5
cycler==0.11.0
decorator==5.1.1
executing==1.1.0
Flask==2.2.2
fonttools==4.37.4
geos==0.2.3
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
numpy==1.23.3
packaging==21.3
pandas==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
pyproj==3.4.0
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2022.4
rasterio==1.3.2
Shapely==1.8.4
six==1.16.0
snuggs==1.4.7
stack-data==0.5.1
traitlets==5.4.0
wcwidth==0.2.5
Werkzeug==2.2.2
xarray==2022.9.0
